feat: Commodities can be input in the web page
I used AutoComplete for this, because Chips and InputChips are both deprecated.

Anyway, I noticed that we use <Form> at some places, but that's not PrimeVue's form (and it shouldn't be) since we don't register it in main.ts. When I tried switching it to <form>, login didn't work. So I'll just leave it as is and pray everything works.

The reason I noticed in the first place is that I get warnings in the console when I open the RoomCreate page.