feat: Availability Calendar properly handles overlapping ranges
The one with the most narrow range wins. It's not perfect logic, but otherwise we'd have to validate server side to not allow specific kinds of intersections which is too much work. Our system is powerful enough as-is. If we encounter a major bug in the logic, I can fix it.